Control.Tracer.Transformers.ObserveOutcome

Description

Observing events with annotations of thread id and time.

transformer

class Monad m => Outcome m a where

Associated Types

type IntermediateValue a

type OutcomeMetric a

Methods

classifyObservable :: a -> m OutcomeProgressionStatus

captureObservableValue :: a -> m (IntermediateValue a)

computeOutcomeMetric :: a -> IntermediateValue a -> IntermediateValue a -> m (OutcomeMetric a)

data OutcomeProgressionStatus

Constructors

OutcomeStarts

OutcomeOther

OutcomeEnds

mkOutcomeExtractor :: (MonadIO m, Outcome m a) => m (OutcomeEnhancedTracer m a -> Tracer m a)

Generic Trace transformer (reactive). It could be written to take an initial argument, but restricting the scope of that per-invocation state seems more appropriate (for the moment). That may be of use if/when explict management of timeout was required and/or non-termination of the outcome at the end of a run was of interest.